+++ spec/dummy/config/routes.rb
@@ -4,6 +4,7 @@ Rails.application.routes.draw do
     post :rename, on: :collection
     post :categorize, on: :collection
     post :copy, on: :collection
+    post :test_passthrough, on: :collection
   end
 
   root to: 'topics#new'